# [WIP] Add "tasks-hash" option for join/merge routines