This is also motivated by the deep-copy overhead discussed in #5739 

`rearrange_by_column_tasks` is already capable of using hashing to split dataframe partitions (avoiding the need to assign a temporary `"_partitions"` column).  This PR just exposes this logic to the `merge`/`join` routines by allowing users to pass `shuffle="tasks-hash"`. 

This change provides the same performance improvement for [the cudf merge benchmark](https://github.com/rapidsai/dask-cuda/blob/branch-0.12/dask_cuda/benchmarks/local_cudf_merge.py) as #5740 (which may have "correctness" problems):


**AFTER Changes**:
```
Merge benchmark
--------------------------
Chunk-size  | 100000000
Frac-match  | 0.3
Ignore-size | 1.05 MB
Protocol    | ucx
Device(s)   | 0,1,2,3
rmm-pool    | True
tcp         | True
ib          | False
nvlink      | True
==========================
Total time  | 716.79 ms
Total time  | 709.37 ms
Total time  | 673.95 ms
==========================

```
<details>
<summary>Worker-Worker Transfer Rates</summary>

```
(w1,w2)     | 25% 50% 75% (total nbytes)
--------------------------
(00,01)     | 4.63 GB/s 5.53 GB/s 7.03 GB/s (11.40 GB)
(00,02)     | 4.91 GB/s 5.16 GB/s 5.52 GB/s (9.00 GB)
(00,03)     | 5.22 GB/s 5.77 GB/s 6.14 GB/s (6.60 GB)
(01,00)     | 4.38 GB/s 5.20 GB/s 6.68 GB/s (11.40 GB)
(01,02)     | 5.04 GB/s 5.72 GB/s 5.80 GB/s (6.60 GB)
(01,03)     | 4.33 GB/s 5.36 GB/s 6.78 GB/s (9.00 GB)
(02,00)     | 5.33 GB/s 5.60 GB/s 7.13 GB/s (9.00 GB)
(02,01)     | 5.87 GB/s 6.19 GB/s 7.02 GB/s (6.60 GB)
(02,03)     | 5.78 GB/s 5.94 GB/s 6.39 GB/s (9.00 GB)
(03,00)     | 5.33 GB/s 5.78 GB/s 6.13 GB/s (6.60 GB)
(03,01)     | 4.41 GB/s 6.04 GB/s 6.70 GB/s (9.00 GB)
(03,02)     | 5.41 GB/s 5.69 GB/s 6.07 GB/s (9.00 GB)
```
</details>


**BEFORE Changes**:
```
Merge benchmark
--------------------------
Chunk-size  | 100000000
Frac-match  | 0.3
Ignore-size | 1.05 MB
Protocol    | ucx
Device(s)   | 0,1,2,3
rmm-pool    | True
tcp         | True
ib          | False
nvlink      | True
==========================
Total time  | 1.59 s
Total time  | 1.85 s
Total time  | 1.79 s
==========================

```
<details>
<summary>Worker-Worker Transfer Rates</summary>

```
(w1,w2)     | 25% 50% 75% (total nbytes)
--------------------------
(00,01)     | 3.36 GB/s 4.20 GB/s 5.54 GB/s (6.50 GB)
(00,02)     | 4.16 GB/s 4.55 GB/s 4.84 GB/s (8.90 GB)
(00,03)     | 3.97 GB/s 5.10 GB/s 6.93 GB/s (11.30 GB)
(01,00)     | 3.74 GB/s 4.15 GB/s 5.54 GB/s (7.20 GB)
(01,02)     | 3.75 GB/s 4.73 GB/s 6.42 GB/s (10.30 GB)
(01,03)     | 4.30 GB/s 5.22 GB/s 6.56 GB/s (11.00 GB)
(02,00)     | 4.17 GB/s 4.76 GB/s 6.55 GB/s (9.60 GB)
(02,01)     | 4.78 GB/s 5.86 GB/s 6.68 GB/s (12.00 GB)
(02,03)     | 4.92 GB/s 5.76 GB/s 6.14 GB/s (7.20 GB)
(03,00)     | 3.60 GB/s 5.56 GB/s 7.38 GB/s (10.30 GB)
(03,01)     | 3.50 GB/s 4.68 GB/s 5.75 GB/s (9.60 GB)
(03,02)     | 4.39 GB/s 4.86 GB/s 6.59 GB/s (10.30 GB)
```
</details>


- [ ] Tests added / passed
- [ ] Passes `black dask` / `flake8 dask`