@rjzamora 2019-12-21
> Adding a new shuffle mechanism seems very expensive to me from a maintenance perspective (all dataframe maintainers will have to be educated about why this is important eventually). If we can get #5740 to work instead then that would be great.

Completely agree - Just wanted to open a draft pr to document this as a possible alternative.