@rjzamora 2020-04-02
@mrocklin - We currently have a [`rearrange_by_hash`](https://github.com/rapidsai/cudf/blob/c24e65218429be7ebf64069b10e81a42117bb73d/python/dask_cudf/dask_cudf/sorting.py#L39) shuffle mechanism in `dask_cudf`.  The *public* dask_cudf api using this mechenism is `DataFrame.repartition()` (i.e. a new `columns=` argument can be used to specify the list of columns for hash-based repartitioning).  This *repartitioning* approach does not result in an ordered index, and thus does not result in known divisions.  However, we have found this algorithm to be more-or-less indispensable for certain workloads.

In order to reduce memory pressure (a serious concern for us), we explicitly avoid the creation of the `"_partitions"` column within `rearrange_by_hash`.  In this sense, we are essentially doing the same thing as the `shuffle="task-hash"` code path added in this PR.

The reason I am bringing this feature up here is: (1) our long-term goal is to get the (multi-column) hash-based repartitioning feature moved into up-stream dask, and (2) we are very motivated to modify the up-stream join/merge algorithm to avoid the creation of the `"_partitions"` column.  Our reasoning for (2) is no longer the deep-copy overhead (which has been addressed), but the fact that the need to carry around an extra column incurs significant memory/communication overhead at scale. 

My original plan was to implement something like `rearrange_by_hash` in `dask.dataframe`, and to simply change the `hash_join` code to use this new algorithm (rather than `shuffle.shuffle`).  However, I am a bit fearful that there will be resistence to the addition of new logic in `dask.dataframe.shuffle` - Do you have any thoughts/advice on the best path forward?

cc @VibhuJawa @randerzander