@TomAugspurger 2020-04-02
Thanks for the status update.

FWIW, I would personally be interested in (optionally) decoupling partitioning from the index. I could imagine an update to the dask.dataframe data model to support (hash) partitioning on a column. This would let us optimize things like boolean filters on the partition column

```python
df2 = df.repartition(column="id")
subset = df2[df2.id == "A"]  # only needs to scan partitions where `id == A`.
```

While updating the data model like this is a big project, a function like `rearrange_by_hash` is a (necessary?) step towards it. Hopefully that goal would be enough to assuage concerns about additional logic in `shuffle` :)