@rjzamora 2020-04-03
Thanks for the feedback @TomAugspurger!

I will submit a draft PR with some minor changes to the shuffle code.  The existing `rearrange_by_column_tasks` can (almost) already handle hash-based repartitioning with/without multiple index columns, so there is no need to add a new `rearrange_by_hash` function.  Instead, the `shuffle_group` and `shuffle_group_2` utility functions can be expanded to perform hashing on demand (when the shuffle `index` is 1+ columns that are **not** `"_partitions"`).