@mrocklin 2020-04-03
> we are very motivated to modify the up-stream join/merge algorithm to avoid the creation of the "_partitions" column. Our reasoning for (2) is no longer the deep-copy overhead (which has been addressed), but the fact that the need to carry around an extra column incurs significant memory/communication overhead at scale

Would this be mostly resolved by making `_partitions` a smaller dtype?  I think that we currently use `uint64` by default, but if we were careful we could shrink this down to `uint8` or `uint16` most of the time, which would likely reduce the scale of this problem significantly.  We used to do this, but ran into corner cases and so we just defaulted to a larger dtype.  If folks are around to identify and work through the corner cases though then this might be a decent solution?  if you have less than 256 partitions then we can shrink the memory use by 8x.