@mrocklin 2020-04-03
> I am a bit fearful that there will be resistence to the addition of new logic in dask.dataframe.shuffle 

Without thinking too much about this (sorry, I think it'll probably take me a while to come up to speed, and I'm short on time these days) I think that additions of large complex bodies of code to the shuffle codebase would need to be backed up with large improvements to general and common workflows.  

Submitting a PR is always welcome because it makes it more clear what the long term cost is.  I'm sorry that I'm unable to give a sense for whether this will likely be accepted ahead of time.