@mrocklin 2020-04-03
@TomAugspurger I'm not sure I understand how this is a step towards separating partitioning from the index.