@rjzamora 2020-04-03
> Would this be mostly resolved by making _partitions a smaller dtype? I think that we currently use uint64 by default, but if we were careful we could shrink this down to uint8 or uint16 most of the time, which would likely reduce the scale of this problem significantly. We used to do this, but ran into corner cases and so we just defaulted to a larger dtype. If folks are around to identify and work through the corner cases though then this might be a decent solution? if you have less than 256 partitions then we can shrink the memory use by 8x.

Good suggestion - This certainly helps, but I had trouble going below int32 in the past.  I'll experiment with this again and collect performance numbers to compare with the "_partitions"-free approach.

> To be clear, when you say "repartitioning" do you mean that you're just moving the boundaries of the partitions around, but not changing the order of rows, as is done in the repartition function, or are you moving around all of the rows so that things are grouped together with related rows. If so, I'm going to suggest that we call this shuffling rather than repartitioning to avoid a naming collision within dask.dataframe nomenclature.

Yes - I am not particularly attached to this functionality being exposed within the repartitioing API. We are **not** repartitioning in the same sense as the other Dask options.  Rather, we are shuffling the data to get unique keys, for a specified set of columns, into the same partition.  The inclusion of this feature as a `columns=` argument to `repartition` was actually chosen to align with a very similar naming/functionallity in Spark.  There are dask_cudf users who like this name/api, but I doubt this is a sticking point on our end.

> @TomAugspurger I'm not sure I understand how this is a step towards separating partitioning from the index.

I believe Tom is generally refering to the effort to implement a public api for the shuffling/paritioning of DataFrames by non-index columns.  The key functionality is more-or-less already in place, the bigger question is really "How should we expose this functionality to the user?"